refactor(#741): refactor regionController
