+++ api/controllers/v1/region/find-all.js
@@ -1,0 +1,14 @@
+const ControllerService = require('../../../services/ControllerService');
+
+module.exports = (req, res) => {
+  TRegion.find().exec((err, found) => {
+    const params = {
+      controllerMethod: 'TRegionController.findAll',
+      searchedItem: 'All regions',
+    };
+    const formattedFound = {
+      regions: found,
+    };
+    return ControllerService.treat(req, err, formattedFound, params, res);
+  });
+};
